refac: move admin functions to specific module
